Web Development Project
Bound Art Book Fair. It is hard to remember where I saw a specific book and retain all the information about the publisher and artist. It's hard to decide which title to buy as funds are limited. I would like to build an app that helps me scan titles easily put them in a basket or wish list to save them for later.